final tidy & reorg
