+++ src/components/SettingsPanel/SettingsPanel.js
@@ -4,7 +4,8 @@ import Transition from 'react-transition-group/Transition';
 import debounce from 'lodash.debounce';
 import { MIN_FONT_SIZE, MAX_FONT_SIZE, COLOR_UPDATE_DELAY } from 'config.js';
 import { getAutoTextColor } from 'utils/colour';
-import { hasDefaultParams, getQueryParams } from 'utils/url';
+import { getQueryParams } from 'utils/url';
+import { hasDefaultParams } from 'utils/parse';
 import { DownArrowIcon, CloseIcon } from 'shared/icons';
 import { Range } from 'components/Range/Range.styles';
 import { ColorPicker } from 'components/ColorPicker/ColorPicker';
